% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{az_line_area}
\alias{az_line_area}
\title{line areas calcualtion for azure}
\usage{
az_line_area(line)
}
\arguments{
\item{line}{line obj returnef from azure api}
}
\description{
line areas calcualtion for azure
}
