% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_result_utls.R
\name{az_lines_to_df}
\alias{az_lines_to_df}
\title{convert azure lines to df}
\usage{
az_lines_to_df(lines)
}
\arguments{
\item{lines}{result lines}
}
\description{
convert azure lines to df
}
