% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_azure_with_cv2.R
\name{az_to_cv2_box}
\alias{az_to_cv2_box}
\title{azure lines into cv2 bounding boxes}
\usage{
az_to_cv2_box(bounds_df, res_lines)
}
\arguments{
\item{bounds_df}{bounds_df result from crop_out_boxes function}

\item{res_lines}{lines result from the azure api reading the whole page}
}
\description{
azure lines into cv2 bounding boxes
}
