% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_result_utls.R
\name{az_words_to_df}
\alias{az_words_to_df}
\title{convert azure line words to df}
\usage{
az_words_to_df(line, type = "word")
}
\arguments{
\item{line}{line of azure result}

\item{type}{word or line}
}
\description{
convert azure line words to df
}
