% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_vis.R
\name{azure_get}
\alias{azure_get}
\title{azure get}
\usage{
azure_get(response, headers)
}
\arguments{
\item{response}{result of azure post}

\item{headers}{also result of azure post}
}
\description{
azure get
}
