% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_vis.R
\name{azure_post}
\alias{azure_post}
\title{azure post}
\usage{
azure_post(subscription_key, endpoint, image_path)
}
\arguments{
\item{subscription_key}{azure key}

\item{endpoint}{azure service end point}

\item{image_path}{path of image to recognize}
}
\description{
azure post
}
