% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure_vis.R
\name{azure_vis}
\alias{azure_vis}
\title{azure vision call function}
\usage{
azure_vis(subscription_key, endpoint, image_path)
}
\arguments{
\item{subscription_key}{azure key}

\item{endpoint}{azure service end point}

\item{image_path}{path of image to recognize}
}
\description{
azure vision call function
}
