% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{bbox_df_to_c}
\alias{bbox_df_to_c}
\title{convert bbox df format to vector format}
\usage{
bbox_df_to_c(df)
}
\arguments{
\item{df}{bbox df}
}
\description{
convert bbox df format to vector format
}
