% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{boundingRect}
\alias{boundingRect}
\title{get bounding box of contour matrix}
\usage{
boundingRect(cnt_mat, img_max_y)
}
\arguments{
\item{cnt_mat}{contour matrix}

\item{img_max_y}{the max y variable of image}
}
\description{
get bounding box of contour matrix
}
