% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy response R functions.R
\name{bounds_to_df}
\alias{bounds_to_df}
\title{python bounds to df}
\usage{
bounds_to_df(bounds, feature)
}
\arguments{
\item{bounds}{python bounds option}

\item{feature}{what type of features, py$FEATURE class}
}
\description{
convert python bounds object to tidy dataframe
}
