% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{chk_box_in}
\alias{chk_box_in}
\title{check whether bounding box is in another bounding box}
\usage{
chk_box_in(box1, box2, err = 5)
}
\arguments{
\item{box1}{first box}

\item{box2}{second box}

\item{err}{fuzzy around several ptx}
}
\description{
check whether bounding box is in another bounding box
}
