% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-process.R
\name{crop_out_boxes}
\alias{crop_out_boxes}
\title{crop out boxes}
\usage{
crop_out_boxes(img_file, hmax)
}
\arguments{
\item{img_file}{file of image}

\item{hmax}{max height allowed for boxes}
}
\description{
crop out boxes
}
