% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-process.R
\name{crop_out_obj}
\alias{crop_out_obj}
\title{object detection}
\usage{
crop_out_obj(image_file, output_cropped = F, output_dir = NULL)
}
\arguments{
\item{image_file}{read image in}

\item{output_cropped}{whether to output cropped objects}

\item{output_dir}{output cropped object to the path}
}
\description{
object detection
}
