% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_process.R
\name{crt_png_from_pdf}
\alias{crt_png_from_pdf}
\title{create png from pdf}
\usage{
crt_png_from_pdf(pdf_file, pages = NULL, dpi = 350)
}
\arguments{
\item{pdf_file}{pdf file}

\item{pages}{pages to create, default to NULL}

\item{dpi}{dpi, quality of the image being created}
}
\description{
create png from pdf
}
