% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{extend_horizontal_lines}
\alias{extend_horizontal_lines}
\title{extend lines in case lines are broken}
\usage{
extend_horizontal_lines(img)
}
\arguments{
\item{img}{Rvision::image obj}
}
\description{
extend lines in case lines are broken
}
