% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_checkboxes.R
\name{get_chkbox_options}
\alias{get_chkbox_options}
\title{get checkbox options}
\usage{
get_chkbox_options(chkbox_df, words_df, question_df)
}
\arguments{
\item{chkbox_df}{data frame with checkbox information}

\item{words_df}{data frame with all the words info from azure}

\item{question_df}{data frame with all the questions}
}
\description{
get checkbox options
}
