% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_checkboxes.R
\name{get_chkbox_questions}
\alias{get_chkbox_questions}
\title{get checkbox questions}
\usage{
get_chkbox_questions(chkbox_df, lines_df)
}
\arguments{
\item{chkbox_df}{data frame with checkbox information}

\item{lines_df}{data frame with all the lines info from azure}
}
\description{
get checkbox questions
}
