% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_checkboxes.R
\name{get_chkbox_wrapper}
\alias{get_chkbox_wrapper}
\title{get checkboxes questions, options, selections}
\usage{
get_chkbox_wrapper(chkbox_df, words_df, lines_df, img)
}
\arguments{
\item{chkbox_df}{data frame with checkbox information}

\item{words_df}{data frame with all the words info from azure}

\item{lines_df}{data frame with all the lines info from azure}

\item{img}{the image np array of the page}
}
\description{
get checkboxes questions, options, selections
}
