% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_checkboxesR.R
\name{get_chkbox_wrapperR}
\alias{get_chkbox_wrapperR}
\title{get checkboxes questions, options, selections}
\usage{
get_chkbox_wrapperR(chkbox_df, words_df, lines_df, img, cl = 1)
}
\arguments{
\item{chkbox_df}{data frame with checkbox information}

\item{words_df}{data frame with all the words info from azure}

\item{lines_df}{data frame with all the lines info from azure}

\item{img}{the image np array of the page}

\item{cl}{number of cores being used}
}
\description{
get checkboxes questions, options, selections
}
