% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ocr_azure.R
\name{get_cropped_azure}
\alias{get_cropped_azure}
\title{post cropped image to azure}
\usage{
get_cropped_azure(res)
}
\arguments{
\item{df}{the data frame with contour info}
}
\description{
post cropped image to azure
}
