% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ocr_azure.R
\name{get_ocr_azure}
\alias{get_ocr_azure}
\title{get ocr from azure}
\usage{
get_ocr_azure(df, cropped_dir_path, img, azure_creds)
}
\arguments{
\item{df}{the data frame with contour info}

\item{cropped_dir_path}{tmp path to output cropped img}

\item{img}{original image}

\item{azure_creds}{credential for azure app}
}
\description{
get ocr from azure
}
