% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{get_question_df}
\alias{get_question_df}
\title{get tidy questions}
\usage{
get_question_df(results)
}
\arguments{
\item{results}{results from the ocr_pdf function}
}
\description{
get tidy questions
}
