% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_checkboxes.R
\name{identify_chkboxes}
\alias{identify_chkboxes}
\title{get checkboxes}
\usage{
identify_chkboxes(img_file)
}
\arguments{
\item{img_file}{file of image or np array of image}
}
\description{
get checkboxes
}
