% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_checkboxes.R
\name{identify_chkboxes_by_parts}
\alias{identify_chkboxes_by_parts}
\title{get checkboxes by parts}
\usage{
identify_chkboxes_by_parts(bounds_df, color_img)
}
\arguments{
\item{bounds_df}{identify bounds data frame}

\item{color_img}{colored image}
}
\description{
get checkboxes by parts
}
