% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_checkboxesR.R
\name{identify_chkboxes_by_partsR}
\alias{identify_chkboxes_by_partsR}
\title{get checkboxes by parts}
\usage{
identify_chkboxes_by_partsR(bounds_df, removed_img, cl = 1)
}
\arguments{
\item{bounds_df}{identify bounds data frame}

\item{removed_img}{image with hw removed}

\item{cl}{number of cores to use}
}
\description{
get checkboxes by parts
}
