% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_checkboxesR.R
\name{identify_chkboxesR}
\alias{identify_chkboxesR}
\title{get checkboxes}
\usage{
identify_chkboxesR(img_file)
}
\arguments{
\item{img_file}{file of image or np array of image}
}
\description{
get checkboxes
}
