% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-process.R
\name{output_cropped_img}
\alias{output_cropped_img}
\title{output_cropped_img}
\usage{
output_cropped_img(cropped_dir_path, img, idx, x, y, w, h)
}
\arguments{
\item{cropped_dir_path}{temporary path for cropped out img}

\item{img}{original image}

\item{idx}{index of image file}

\item{x}{x coord}

\item{y}{y coord}

\item{w}{width}

\item{h}{height}
}
\description{
output_cropped_img
}
