% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{parse_coord}
\alias{parse_coord}
\title{turning bounding obj to x,y coordinates}
\usage{
parse_coord(x)
}
\arguments{
\item{x}{the bounding obj from python}
}
\description{
turning bounding obj to x,y coordinates
}
