% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdftxt_to_az_format.R
\name{pdf_txt_to_az_df}
\alias{pdf_txt_to_az_df}
\title{pdf_txt_df to az table}
\usage{
pdf_txt_to_az_df(pdf_txt_df)
}
\arguments{
\item{pdf_txt_df}{result of pdf_data function of pdftools}
}
\description{
pdf_txt_df to az table
}
