% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ocr_azure.R
\name{post_cropped_azure}
\alias{post_cropped_azure}
\title{post cropped image to azure}
\usage{
post_cropped_azure(
  df,
  cropped_dir_path,
  img,
  azure_creds,
  box_highlight = T,
  remove_fl = T
)
}
\arguments{
\item{df}{the data frame with contour info}

\item{cropped_dir_path}{tmp path to output cropped img}

\item{img}{original image}

\item{azure_creds}{credential for azure app}

\item{box_highlight}{whether to have the additional step of flagging the cropped image}

\item{remove_fl}{whether to remove images}
}
\description{
post cropped image to azure
}
