% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{pts_to_wh}
\alias{pts_to_wh}
\title{convert point coords to x,y, w, h}
\usage{
pts_to_wh(boundingbox)
}
\arguments{
\item{boundingbox}{the bounding box returned by azure api}
}
\description{
convert point coords to x,y, w, h
}
