% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{ptwise_chk_approx}
\alias{ptwise_chk_approx}
\title{check whether two approx points are really close}
\usage{
ptwise_chk_approx(approx)
}
\arguments{
\item{approx}{approximated vertex points of a contour}
}
\description{
check whether two approx points are really close
}
