% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{quick_img_chkR}
\alias{quick_img_chkR}
\title{quick image chk}
\usage{
quick_img_chkR(df, img, out_fl = "new.png")
}
\arguments{
\item{df}{x, y, w, h df}

\item{img}{raw img}

\item{out_fl}{output file}
}
\description{
quick image chk
}
