% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-process.R
\name{remove_color}
\alias{remove_color}
\title{remove non black and white color from an image to remove handwritten}
\usage{
remove_color(img_file)
}
\arguments{
\item{img_file}{the file path of the image}
}
\description{
remove non black and white color from an image to remove handwritten
}
