% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-processR.R
\name{remove_colorR}
\alias{remove_colorR}
\title{remove non black and white color from an image to remove handwritten}
\usage{
remove_colorR(img_file, min_clr = 100, max_clr = 250)
}
\arguments{
\item{img_file}{the file path of the image}

\item{min_clr}{minimum color number to filter}

\item{max_clr}{maximum color number to filter}
}
\description{
remove non black and white color from an image to remove handwritten
}
