% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_process.R
\name{resize_png}
\alias{resize_png}
\title{resize the image to meet azure criteria}
\usage{
resize_png(img_file, file_size_limit = 3.8)
}
\arguments{
\item{img_file}{original pdf file path}

\item{file_size_limit}{file size limit in mbs}
}
\description{
resize the image to meet azure criteria
}
