% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-process.R
\name{sort_contours}
\alias{sort_contours}
\title{sort contours}
\usage{
sort_contours(cnts, method = "left-to-right", hmax = 100)
}
\arguments{
\item{cnts}{contours}

\item{method}{'left-to-right' or 'bottom-to-top'}

\item{hmax}{max height to include}
}
\description{
sort contours
}
