% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-processR.R
\name{sort_contoursR}
\alias{sort_contoursR}
\title{sort contours}
\usage{
sort_contoursR(cnts, method = "left-to-right", hmax = 100, img_max_y)
}
\arguments{
\item{cnts}{contours}

\item{method}{'left-to-right' or 'bottom-to-top'}

\item{hmax}{max height to include}

\item{img_max_y}{the max y variable of image}
}
\description{
sort contours
}
