% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy response R functions.R
\name{strcture_doc}
\alias{strcture_doc}
\title{python contents object to df}
\usage{
strcture_doc(contents)
}
\arguments{
\item{obj}{python block, paragraph, word level info object}
}
\description{
convert python contents object to tidy dataframe
}
