% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy response R functions.R
\name{strcture_words}
\alias{strcture_words}
\title{python words object to df}
\usage{
strcture_words(obj)
}
\arguments{
\item{obj}{python word level info object}
}
\description{
convert python words object to tidy dataframe
}
