% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytbl_to_r.R
\name{tidytbl_to_r}
\alias{tidytbl_to_r}
\title{tidy result to row and column table}
\usage{
tidytbl_to_r(tidy_res, col_header, bw_scale = 5)
}
\arguments{
\item{tidy_res}{tidied result, could be from azure after tidying or pdf_data}

\item{col_header}{column header to help identifying the tabular structure}

\item{bw_scale}{scaling on height when determine row locations}
}
\description{
tidy result to row and column table
}
