% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ocr_azure.R
\name{vec_get_cropped_azure}
\alias{vec_get_cropped_azure}
\title{get response from azure api, vectorized}
\usage{
vec_get_cropped_azure(df)
}
\arguments{
\item{df}{the data frame from vec_post_cropped_azure}
}
\description{
get response from azure api, vectorized
}
