% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ocr_azure.R
\name{vec_post_cropped_azure}
\alias{vec_post_cropped_azure}
\title{vectorized post azure}
\usage{
vec_post_cropped_azure(
  df,
  cropped_tm_dir,
  img,
  azure_creds,
  push_to_az = T,
  box_highlight = F,
  remove_fl = F
)
}
\arguments{
\item{df}{bounds dataframe}

\item{cropped_tm_dir}{temporary dir to output cropped img}

\item{img}{image}

\item{push_to_az}{whether to push to azure}

\item{box_highlight}{whether to have the additional step of flagging the cropped image}

\item{remove_fl}{whether to remove cropped files}

\item{azure_cred}{azure credentials}
}
\description{
vectorized post azure
}
