% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ocr_azure.R
\name{ocr_img_wrapper}
\alias{ocr_img_wrapper}
\title{ocr image wrapper}
\usage{
ocr_img_wrapper(img_file, hmax = 100, cropped_tm_dir, azure_creds)
}
\arguments{
\item{img_file}{img file path}

\item{hmax}{maximum box height to include}

\item{cropped_tm_dir}{temporary dir for cropped img}

\item{azure_creds}{azure credential}
}
\description{
ocr image wrapper
}
