% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_process.R
\name{resize_png}
\alias{resize_png}
\title{resize the image to meet azure criteria}
\usage{
resize_png(img_file, new_size = "3500x3500")
}
\arguments{
\item{img_file}{original file path}

\item{new_size}{in the form of YxY}
}
\description{
resize the image to meet azure criteria
}
