% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-process.R
\name{sort_contours}
\alias{sort_contours}
\title{sort contours}
\usage{
sort_contours(cnts, method = "left-to-right")
}
\arguments{
\item{cnts}{contours}

\item{method}{'left-to-right' or 'bottom-to-top'}
}
\description{
sort contours
}
